admin-panel/vizual_proqramlasdirma_version2/dao-b1-main/backend/requirements.txt
# Core Backend Dependencies
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
fastapi>=0.95
uvicorn>=0.23
strawberry-graphql>=0.153
slowapi>=0.1.4
websockets>=11.0
pydantic>=2.3
httpx>=0.24
jsonrpcserver>=5.0

# Sprint 3: Data Science & Monitoring Dependencies
# Required for dashboard.py and monitoring.py
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
prometheus-client>=0.17.0

# ML & Analytics (Sprint 3)
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Testing
pytest>=7.4.0
